+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Add LaunchImpl.log(Maybe<java.lang.String>, jFunction<String, SaveLogRQ>) method, by @HardNorth
 
 ## [5.3.11]
 ### Fixed